+++ src/components/skills/absorb_shield.rs
@@ -1,6 +1,6 @@
-use crate::common::Vec2;
-use crate::components::controller::CharEntityId;
-use crate::components::skills::skills::{SkillDef, SkillManifestation, SkillTargetType};
+use crate::components::skills::skills::{
+    FinishCast, SkillDef, SkillManifestation, SkillTargetType,
+};
 use crate::components::status::absorb_shield::AbsorbStatus;
 use crate::components::status::status::ApplyStatusComponent;
 use crate::configs::DevConfig;
@@ -17,11 +17,7 @@ impl SkillDef for AbsorbShieldSkill {
 
     fn finish_cast(
         &self,
-        caster_entity_id: CharEntityId,
-        _caster_pos: Vec2,
-        _skill_pos: Option<Vec2>,
-        _char_to_skill_dir: &Vec2,
-        target_entity: Option<CharEntityId>,
+        params: &FinishCast,
         ecs_world: &mut specs::world::World,
     ) -> Option<Box<dyn SkillManifestation>> {
         let mut sys_vars = ecs_world.write_resource::<SystemVariables>();
@@ -34,9 +30,13 @@ impl SkillDef for AbsorbShieldSkill {
         sys_vars
             .apply_statuses
             .push(ApplyStatusComponent::from_secondary_status(
-                caster_entity_id,
-                target_entity.unwrap(),
-                Box::new(AbsorbStatus::new(caster_entity_id, now, duration_seconds)),
+                params.caster_entity_id,
+                params.target_entity.unwrap(),
+                Box::new(AbsorbStatus::new(
+                    params.caster_entity_id,
+                    now,
+                    duration_seconds,
+                )),
             ));
         None
     }