+++ src/components/skills/firewall.rs
@@ -8,8 +8,8 @@ use crate::common::{rotate_vec2, v2, Vec2, Vec2i};
 use crate::components::char::{ActionPlayMode, CharacterStateComponent, Team};
 use crate::components::controller::CharEntityId;
 use crate::components::skills::skills::{
-    SkillDef, SkillManifestation, SkillManifestationComponent, SkillManifestationUpdateParam,
-    SkillTargetType, Skills,
+    FinishCast, SkillDef, SkillManifestation, SkillManifestationComponent,
+    SkillManifestationUpdateParam, SkillTargetType, Skills,
 };
 use crate::components::{
     ApplyForceComponent, DamageDisplayType, HpModificationRequest, HpModificationType,
@@ -34,19 +34,15 @@ impl SkillDef for FireWallSkill {
 
     fn finish_cast(
         &self,
-        caster_entity_id: CharEntityId,
-        _caster_pos: Vec2,
-        skill_pos: Option<Vec2>,
-        char_to_skill_dir: &Vec2,
-        _target_entity: Option<CharEntityId>,
+        params: &FinishCast,
         ecs_world: &mut specs::world::World,
     ) -> Option<Box<dyn SkillManifestation>> {
         if let Some(caster) = ecs_world
             .read_storage::<CharacterStateComponent>()
-            .get(caster_entity_id.0)
+            .get(params.caster_entity_id.0)
         {
-            let angle_in_rad = char_to_skill_dir.angle(&Vector2::y());
-            let angle_in_rad = if char_to_skill_dir.x > 0.0 {
+            let angle_in_rad = params.char_to_skill_dir.angle(&Vector2::y());
+            let angle_in_rad = if params.char_to_skill_dir.x > 0.0 {
                 angle_in_rad
             } else {
                 -angle_in_rad
@@ -56,13 +52,13 @@ impl SkillDef for FireWallSkill {
             let mut updater = ecs_world.write_resource::<LazyUpdate>();
             let configs = ecs_world.read_resource::<DevConfig>();
             Some(Box::new(PushBackWallSkill::new(
-                caster_entity_id,
+                params.caster_entity_id,
                 caster.team,
                 configs.skills.firewall.damage,
                 configs.skills.firewall.pushback_force,
                 configs.skills.firewall.force_duration_seconds,
                 &mut ecs_world.write_resource::<PhysicEngine>(),
-                &skill_pos.unwrap(),
+                &params.skill_pos.unwrap(),
                 angle_in_rad,
                 sys_vars.time,
                 sys_vars.tick,