+++ src/components/skills/gaz_turret.rs
@@ -1,10 +1,11 @@
-use crate::common::Vec2;
 use crate::components::char::{
     CharOutlook, CharacterEntityBuilder, CharacterStateComponent, NpcComponent, TurretComponent,
     TurretControllerComponent,
 };
 use crate::components::controller::{CharEntityId, ControllerComponent, ControllerEntityId};
-use crate::components::skills::skills::{SkillDef, SkillManifestation, SkillTargetType};
+use crate::components::skills::skills::{
+    FinishCast, SkillDef, SkillManifestation, SkillTargetType,
+};
 use crate::configs::DevConfig;
 use crate::consts::{JobId, MonsterId};
 use crate::runtime_assets::map::{CollisionGroup, PhysicEngine};
@@ -22,16 +23,12 @@ impl SkillDef for GazTurretSkill {
 
     fn finish_cast(
         &self,
-        caster_entity_id: CharEntityId,
-        _caster_pos: Vec2,
-        skill_pos: Option<Vec2>,
-        _char_to_skill_dir: &Vec2,
-        _target_entity: Option<CharEntityId>,
+        params: &FinishCast,
         ecs_world: &mut World,
     ) -> Option<Box<dyn SkillManifestation>> {
         if let Some(caster) = ecs_world
             .read_storage::<CharacterStateComponent>()
-            .get(caster_entity_id.0)
+            .get(params.caster_entity_id.0)
         {
             let entities = &ecs_world.entities();
             let updater = &ecs_world.read_resource::<LazyUpdate>();
@@ -39,9 +36,9 @@ impl SkillDef for GazTurretSkill {
             updater.insert(char_entity_id.0, NpcComponent);
             CharacterEntityBuilder::new(char_entity_id, "turret")
                 .insert_sprite_render_descr_component(updater)
-                .insert_turret_component(caster_entity_id, updater)
+                .insert_turret_component(params.caster_entity_id, updater)
                 .physics(
-                    skill_pos.unwrap(),
+                    params.skill_pos.unwrap(),
                     &mut ecs_world.write_resource::<PhysicEngine>(),
                     |builder| builder.collision_group(CollisionGroup::Turret).circle(1.0),
                 )
@@ -73,23 +70,20 @@ impl SkillDef for GazDestroyTurretSkill {
 
     fn finish_cast(
         &self,
-        caster_entity_id: CharEntityId,
-        _caster_pos: Vec2,
-        _skill_pos: Option<Vec2>,
-        _char_to_skill_dir: &Vec2,
-        target_entity: Option<CharEntityId>,
+        params: &FinishCast,
         ecs_world: &mut World,
     ) -> Option<Box<dyn SkillManifestation>> {
-        if target_entity
+        if params
+            .target_entity
             .and_then(|it| {
                 ecs_world
                     .read_storage::<TurretComponent>()
                     .get(it.0)
-                    .map(|turret| turret.owner_entity_id == caster_entity_id)
+                    .map(|turret| turret.owner_entity_id == params.caster_entity_id)
             })
             .unwrap_or(false)
         {
-            let target_entity = target_entity.unwrap();
+            let target_entity = params.target_entity.unwrap();
             if let Some(turret) = ecs_world
                 .write_storage::<CharacterStateComponent>()
                 .get_mut(target_entity.0)
@@ -116,16 +110,12 @@ impl SkillDef for GazTurretTargetSkill {
 
     fn finish_cast(
         &self,
-        caster_entity_id: CharEntityId,
-        _caster_pos: Vec2,
-        _skill_pos: Option<Vec2>,
-        _char_to_skill_dir: &Vec2,
-        target_entity: Option<CharEntityId>,
+        params: &FinishCast,
         ecs_world: &mut World,
     ) -> Option<Box<dyn SkillManifestation>> {
         for turret in (&mut ecs_world.write_storage::<TurretComponent>()).join() {
-            if turret.owner_entity_id == caster_entity_id {
-                turret.preferred_target = target_entity;
+            if turret.owner_entity_id == params.caster_entity_id {
+                turret.preferred_target = params.target_entity;
             }
         }