+++ src/components/skills/heal.rs
@@ -1,9 +1,9 @@
 use specs::LazyUpdate;
 
-use crate::components::controller::CharEntityId;
-use crate::components::skills::skills::{SkillDef, SkillManifestation, SkillTargetType};
+use crate::components::skills::skills::{
+    FinishCast, SkillDef, SkillManifestation, SkillTargetType,
+};
 
-use crate::common::Vec2;
 use crate::components::{HpModificationRequest, HpModificationType, SoundEffectComponent};
 use crate::configs::DevConfig;
 use crate::systems::SystemVariables;
@@ -19,14 +19,10 @@ impl SkillDef for HealSkill {
 
     fn finish_cast(
         &self,
-        caster_entity_id: CharEntityId,
-        caster_pos: Vec2,
-        _skill_pos: Option<Vec2>,
-        _char_to_skill_dir: &Vec2,
-        target_entity: Option<CharEntityId>,
+        params: &FinishCast,
         ecs_world: &mut specs::world::World,
     ) -> Option<Box<dyn SkillManifestation>> {
-        let target_entity_id = target_entity.unwrap();
+        let target_entity_id = params.target_entity.unwrap();
         let entities = &ecs_world.entities();
         let updater = ecs_world.read_resource::<LazyUpdate>();
         let mut sys_vars = ecs_world.write_resource::<SystemVariables>();
@@ -36,12 +32,12 @@ impl SkillDef for HealSkill {
             SoundEffectComponent {
                 target_entity_id,
                 sound_id: sys_vars.assets.sounds.heal,
-                pos: caster_pos,
+                pos: params.caster_pos,
                 start_time: sys_vars.time,
             },
         );
         sys_vars.hp_mod_requests.push(HpModificationRequest {
-            src_entity: caster_entity_id,
+            src_entity: params.caster_entity_id,
             dst_entity: target_entity_id,
             typ: HpModificationType::Heal(ecs_world.read_resource::<DevConfig>().skills.heal.heal),
         });